added springfox for swagger 2 and swagger ui, added repositories and services for timetables and purchases, replaced method getting user id on UserUtil, and other

+++ src/main/java/ru/senla/realestatemarket/model/property/ApartmentProperty.java
@@ -32,7 +32,7 @@ public class ApartmentProperty extends HousingProperty {
     @Column(name = "apartment_number")
     private String apartmentNumber;
 
-    private Integer floor;
+    private Short floor;
 
 
     @PostConstruct
